Authentication/requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
sqlalchemy==1.4.32
psycopg2==2.9.1
python-jose==3.3.0
